+++ packages/python_ffi_cpython_dart/lib/src/extensions/convert_to_dart_extension.dart
@@ -43,6 +43,9 @@ extension _ConvertToDartExtension on Pointer<PyObject> {
     if (isSet(platform)) {
       return asSet(platform);
     }
+    if (isClassDefinition(platform)) {
+      return _PythonClassDefinitionCPython(platform, object);
+    }
     if (isIterator(platform)) {
       return asIterator(platform);
     }